Migrate from tslint to eslint with typescript
This was not an easy migration, not the least because the automatic
conversion script failed to work for me.  But I built a new eslintrc.js
from scratch and am reasonably happy with it so there's that.

I've tried to fix as many lint issues as I can but there's something
weird about the AWS Amplify stuff so I'm just leaving a couple errors in
for now until I can figure out how to silence them or fix them.  I
verified that the pages in question still function as expected.